hello world

+++ 0x00-hello_world/6-size.c
@@ -1,11 +1,9 @@
 #include <stdio.h>
 
 /**
-* main - print the string in the put function
+* main - A program that prints the size of various computer types
 *
-* Description: using the main function
-* this program prints "Programming is like building a multilingual puzzle
-* Return: 0
+* Return 0 (Success)
 */
 int main(void)
 {
@@ -15,10 +13,10 @@ int main(void)
 	long long lli;
 	float f;
 
-	printf("Size of a char: %ld byte(s)\n", sizeof(c));
-	printf("Size of a int: %ld byte(s)\n", sizeof(i));
-	printf("Size of a long: %ld byte(s)\n", sizeof(li));
-	printf("Size of a long long: %ld byte(s)\n", sizeof(lli));
-	printf("Size of a float: %ld byte(s)\n", sizeof(f));
-	return (0);
+printf("Size of a char: %lu byte(s)\n",(unsigned long)sizeof(c));
+printf("Size of a int: %lu byte(s)\n",(unsigned long)sizeof(i));
+printf("Size of a long: %lu byte(s)\n",(unsigned long)sizeof(li));
+printf("Size of a long long: %lu byte(s)\n",(unsigned long)sizeof(lli));
+printf("Size of a float: %lu byte(s)\n",(unsigned long)sizeof(f));
+return (0);
 }
